chore: renamed all components and folders from PascalCase to kebab-case
